… types/record_storage.go