* Moved most struct allocations from heap to stack
* Removed extra goroutines and channels
* Removed duplicate heap allocations on GenProxy Service

Performance of GenProxy improved from
```
Running 10s test @ http://localhost:80
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    60.30ms  172.49ms   1.06s    91.44%
    Req/Sec   811.14    228.33     1.20k    77.06%
  13912 requests in 10.04s, 1.70MB read
Requests/sec:   1385.79
Transfer/sec:    173.22KB
```
to
```
Running 10s test @ http://localhost:80
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.88ms    3.47ms  64.42ms   95.03%
    Req/Sec     1.08k   116.70     1.27k    76.50%
  21485 requests in 10.00s, 2.62MB read
Requests/sec:   2147.85
Transfer/sec:    268.48KB
```